+++ SampleLibrary/SampleLibraryInclude.h
@@ -24,4 +24,4 @@
 // Sample C Exports
 
 SAMPLELIBRARY_C_FUNCTION
-double Add(double a, double b);
+double Add(int a, double b);
